# gateway/internal/gateway/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 多平台音乐网关 API
  version: 0.5.3
  description: |
    **多平台音乐网关 API** 是一组用于聚合 QQ 音乐 / 网易云音乐 / 酷狗音乐等多家音乐源的统一 RESTful 接口。
    网关在内部完成账号池调度、限流配额、计费扣减、Extra Usage 超额计费、余额扣款等控制逻辑，
    上游应用只需要通过统一的 API 接入即可。

    ## 协议总览
    - **基础路径**：所有业务接口位于 `/v1/...` 命名空间下。
    - **认证方式**：
      - **Cookie 会话**（前端控制台用）：登录后服务器下发 `sid` 会话 Cookie。
      - **API Key**（程序化调用用）：在 `Authorization: Bearer <api_key>` 请求头中携带。
    - **响应封装**：网关接口绝大多数返回 `{ code, message, data }` 包装结构（`code = 0` 表示成功），
      详情见下方 `Response` schema 定义。
    - **CSRF 保护**：使用 Cookie 会话时，所有非 GET/HEAD/OPTIONS 请求都需要在请求头中携带
      `X-CSRF-Token`（值与 Cookie `csrf_token` 一致）。API Key 模式不受 CSRF 限制。
    - **限流模型**：
      - **QPS**：按用户套餐限制每秒请求数。
      - **每日 / 每月配额**：免费版默认 100 次/日；超出后若开启 Extra Usage 则会按千次价从余额中扣费。

    ## 错误码
    全局错误码定义在 `ErrorCode` 字段：
    | code | 含义 |
    | ---- | ---- |
    | 0 | 成功 |
    | 400 | 请求参数错误 |
    | 401 | 未认证 / 凭据失效 |
    | 403 | 权限不足 |
    | 404 | 资源不存在 |
    | 409 | 资源冲突（重复创建、版本冲突等） |
    | 429 | 限流（QPS 或日/月配额耗尽） |
    | 500 | 服务器内部错误 |
  contact:
    name: 多平台音乐网关
    url: https://github.com/luoye663/qqmusicapi
servers:
  - url: /
    description: 当前实例
  - url: https://api.example.com
    description: 生产环境（占位，按部署实际域名替换）
...
security:
  - cookieAuth: []
  - apiKeyAuth: []
paths:
...
  /v1/billing/plans/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    put:
      tags: [套餐]
      summary: 更新指定套餐（管理员）
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Plan" }
      responses:
        "200":
          description: 更新成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlanEnvelope" }
    delete:
      tags: [套餐]
      summary: 删除套餐（管理员）
      responses:
        "200":
          description: 已删除
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EmptyEnvelope" }
...
  /v1/billing/orders:
    get:
      tags: [订单]
      summary: 列出当前用户订单
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        "200":
          description: 订单列表
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderListEnvelope" }
    post:
      tags: [订单]
      summary: 创建订单（订阅 / 充值）
      description: |
        - 订阅：传 `planId`，订单类型为 `subscription`，支付成功后会升级用户套餐并将其下所有 API Key 切换到该套餐。
        - 充值：传 `topupAmountCents`，订单类型为 `topup`，支付成功后金额自动入账到用户钱包。
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderRequest" }
      responses:
        "200":
          description: 订单已创建，含支付链接
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateOrderResponseEnvelope" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /v1/billing/orders/{id}:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    get:
      tags: [订单]
      summary: 查询订单
      responses:
        "200":
          description: 订单详情
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderEnvelope" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/billing/orders/{id}/refund:
    parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
    post:
      tags: [订单]
      summary: 退款（管理员）
      description: 仅 `admin` / `super_admin` 角色可用。
      responses:
        "200":
          description: 退款发起成功
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderEnvelope" }
        "403": { $ref: "#/components/responses/Forbidden" }

  # ========== 余额 ==========
...
  /v1/admin/settings/payment:
    get:
      tags: [系统设置（管理员）]
      summary: 查询支付配置
      responses:
        "200":
          description: 配置
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentConfigEnvelope" }
    put:
      tags: [系统设置（管理员）]
      summary: 更新支付配置
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PaymentConfig" }
      responses:
        "200":
          description: 已保存
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentConfigEnvelope" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sid
      description: 通过 `/v1/auth/login` 颁发的会话 Cookie，浏览器自动携带；同源调用无需手动配置。
    apiKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: "在 `Authorization: Bearer <api_key>` 中携带。"
...
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
    Forbidden:
      description: 权限不足
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorEnvelope" }
...
  schemas:
...
    Response:
      type: object
      required: [code, message]
      properties:
        code: { type: integer, example: 0, description: "0=成功；非 0 见错误码表" }
        message: { type: string, example: "ok" }
        data: { description: "实际业务数据，类型见各接口" }
    ErrorEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          example: { code: 400, message: "invalid body", data: null }
    EmptyEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { type: object, additionalProperties: true }
...
    Plan:
      type: object
      properties:
        id: { type: string, example: "enterprise" }
        code: { type: string }
        name: { type: string, example: "企业版" }
        priceCents: { type: integer, format: int64 }
        currency: { type: string, default: "CNY" }
        period: { type: string, enum: [monthly, yearly, none] }
        qps: { type: integer }
        dailyLimit: { type: integer, format: int64 }
        monthlyLimit: { type: integer, format: int64 }
        softLimitPct: { type: number, format: float, description: "0-1 之间的预警阈值" }
        payAsYouGo: { type: boolean }
        overagePricePer_1k: { type: number, format: float, description: "超额单价：单位元/千次" }
    PlanEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Plan" }
...
    Order:
      type: object
      properties:
        id: { type: string }
        userId: { type: string }
        planId: { type: string, nullable: true }
        purpose: { type: string, enum: [subscription, topup] }
        amount: { type: string, description: "金额字符串，单位元" }
        currency: { type: string }
        status: { type: string, enum: [pending, paid, refunded, failed, canceled] }
        paymentProvider: { type: string }
        payUrl: { type: string }
        paidAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
    OrderEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Order" }
    OrderListEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                items: { type: array, items: { $ref: "#/components/schemas/Order" } }
                total: { type: integer }
    CreateOrderRequest:
      type: object
      properties:
        planId: { type: string, description: 套餐订阅时使用 }
        topupAmountCents: { type: integer, format: int64, description: 充值时使用，单位分 }
        paymentProvider: { type: string, enum: [yipay, hupijiao], default: yipay }
        successUrl: { type: string, format: uri, description: 用户支付完成跳转地址 }
    CreateOrderResponseEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data:
              type: object
              properties:
                orderId: { type: string }
                payUrl: { type: string, format: uri }
                amount: { type: string }
                currency: { type: string }
                status: { type: string }
                planName: { type: string, nullable: true }
...
    PaymentConfig:
      type: object
      properties:
        yipay:
          type: object
          properties:
            enabled: { type: boolean }
            pid: { type: string }
            key: { type: string }
            apiUrl: { type: string }
        hupijiao:
          type: object
          properties:
            enabled: { type: boolean }
            appId: { type: string }
            appSecret: { type: string }
        defaultProvider: { type: string, enum: [yipay, hupijiao] }
        baseUrl: { type: string, description: "回调地址前缀" }
    PaymentConfigEnvelope:
      allOf:
        - $ref: "#/components/schemas/Response"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/PaymentConfig" }
...